requirements.txt
pandas==2.2.2
requests==2.32.3
aiohttp==3.9.5
python-dotenv==1.0.1
sqlalchemy==2.0.31
plotly==5.22.0
asyncpg==0.29.0
greenlet==3.0.3